Fixed 64-bit template typecast errors
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@7387 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/G2Neighbour.cpp
@@ -556,7 +556,7 @@ BOOL CG2Neighbour::OnPing(CG2Packet* pPacket, BOOL bTCP)
 
 		for ( INT_PTR nCur = 0; (nCur < nCount && nCur < nRelayTo); nCur++ )
 		{
-			INT_PTR nRand = GetRandomNum( 0, pG2Nodes.GetCount() - 1 );
+			INT_PTR nRand = GetRandomNum( 0, (int)pG2Nodes.GetCount() - 1 );
 
 			CG2Neighbour* pNeighbour = pG2Nodes.GetAt( nRand );
 			pNeighbour->Send( pPacket, FALSE );
